api/requirements.txt
fastapi
uvicorn[standard]
supabase==2.0.0
pytest==7.4.3
httpx>=0.24.0,<0.25.0
python-dotenv